pyproject.toml
[tool.poetry]
name = "emart"
version = "0.1.0"
description = ""
authors = ["Ankush <ankush357159@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
python-dotenv = "^0.19.2"
django-cors-headers = "^3.11.0"
cloudinary = "^1.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
